feat(web): consolidate contract address definitions into single configuration source
