+++ apps/web/src/lib/constants.ts
@@ -1,7 +1,5 @@
 import { env } from "./env";
 
-export const PAYMENT_STREAM_CONTRACT_ID = env.NEXT_PUBLIC_PAYMENT_STREAM_CONTRACT_ID;
-export const DISTRIBUTOR_CONTRACT_ID = env.NEXT_PUBLIC_DISTRIBUTOR_CONTRACT_ID;
 export const SOROBAN_RPC_URL = env.NEXT_PUBLIC_SOROBAN_RPC_URL;
 export const NETWORK_PASSPHRASE = env.NEXT_PUBLIC_NETWORK_PASSPHRASE;
 export const STELLAR_NETWORK = env.NEXT_PUBLIC_STELLAR_NETWORK;
