+++ apps/web/src/lib/env.ts
@@ -114,3 +114,6 @@ export const env = process.env.NODE_ENV === 'test'
       NEXT_PUBLIC_STELLAR_HORIZON_URL: process.env.NEXT_PUBLIC_STELLAR_HORIZON_URL ?? 'https://horizon-testnet.stellar.org',
     } as ReturnType<typeof validateEnv>)
   : validateEnv();
+
+export const PAYMENT_STREAM_CONTRACT_ID = env.NEXT_PUBLIC_PAYMENT_STREAM_CONTRACT_ID;
+export const DISTRIBUTOR_CONTRACT_ID = env.NEXT_PUBLIC_DISTRIBUTOR_CONTRACT_ID;
